synetec mvc assessment

+++ Solution/SynetecMvcAssessment/Models/BonusPoolCalculatorModel.cs
@@ -7,8 +7,7 @@ namespace InterviewTestTemplatev2.Models
 {
     public class BonusPoolCalculatorModel
     {
-
-        public int BonusPoolAmount { get; set; }
+        public decimal BonusPoolAmount { get; set; }
         public List<Data.HrEmployee> AllEmployees { get; set; }        
         public int SelectedEmployeeId { get; set; }
